refactor(zama-fhe): tables that tally their own growth, a per-step heap gate, ops split from admission
Answers the structural review on #3502:

- heap_tally.rs owns the allocator growth model and TalliedVec<T>, a
  crate-private vector that records every byte it requests as it grows.
  Growth accounting becomes a property of the container instead of a
  `&mut usize` threaded through builder.rs, lower.rs, and accounts.rs;
  accounts.rs no longer imports anything cost-related.
- builder.rs is only the admission machine: the tables, commit_step, and
  finish. commit_step is the single home of the TooManySteps check and
  the new per-step build-heap gate: a step whose tables push the running
  build tally past the budget is rolled back and rejected at that step,
  so the builder's own allocations can never exhaust the real region
  mid-build. The named op constructors move to ops.rs; the
  lowering/rollback tally protocol moves to lower.rs.
- cost.rs keeps only the ceilings, the trace floor, and the cost DTO, so
  apps importing cost numbers no longer pull in the Vec growth model.
- binding() is now one clone into into_binding(), so the two spellings
  cannot validate or bind differently.
- heap_budget.rs pins the builder side of the host-frame gap:
  the_builder_admits_what_the_host_heap_cannot_hold proves build() Ok at
  15, 16, and 20 shared-audience eight-subject public creates. Both
  counting-allocator proofs stay byte-exact; the frontier is still 43
  admitted of 122.

Two boundary-sweep shapes whose build-alone term crosses the budget now
reject with ExceedsBuildHeapBudget at the crossing step instead of the
finish-time packet error; the packet ceiling stays pinned by a
shared-audience shape that is packet-heavy but build-cheap.
