+++ solana/crates/zama-fhe/src/cost.rs
@@ -82,102 +82,6 @@ pub const CPI_INSTRUCTION_DATA_LIMIT: usize = 10 * 1024;
 /// rent exemption, so counting three never under-counts.
 pub const CPIS_PER_PERSISTENT_CREATE: usize = 3;
 
-/// `RawVec`'s first non-zero capacity for an element size — the other half of the growth model
-/// [`tally_push`] and [`invoke_table_heap_bytes`] share.
-fn minimum_nonzero_capacity(elem_size: usize) -> usize {
-    if elem_size == 1 {
-        8
-    } else if elem_size <= 1024 {
-        4
-    } else {
-        1
-    }
-}
-
-/// The capacity a full `Vec` grows to on the next push: double, or `RawVec`'s minimum first
-/// capacity.
-fn grown_capacity(capacity: usize, elem_size: usize) -> usize {
-    std::cmp::max(2 * capacity, minimum_nonzero_capacity(elem_size))
-}
-
-/// Tallies the bytes the next `push` will request from the allocator, modeling `Vec` growth
-/// the way the never-freeing bump region pays for it: a full vector reallocates to double its
-/// capacity (or to `RawVec`'s minimum first capacity), and the outgrown buffer is never
-/// reclaimed. Call immediately before the push. Validated byte-for-byte against a counting
-/// allocator by `heap_budget.rs`, which is what keeps this model honest if `Vec`'s growth
-/// strategy ever changes.
-pub(crate) fn tally_push<T>(vec: &Vec<T>, tally: &mut usize) {
-    if vec.len() < vec.capacity() {
-        return;
-    }
-    let elem_size = std::mem::size_of::<T>();
-    *tally += grown_capacity(vec.capacity(), elem_size) * elem_size;
-}
-
-/// Bytes `count` pushes into an empty `Vec` request from the allocator, and the capacity the
-/// vector ends at — the growth pattern Anchor's generated `to_account_metas`/`to_account_infos`
-/// pay for the fixed CPI accounts, which start from `vec![]`.
-fn pushes_from_empty(count: usize, elem_size: usize) -> (usize, usize) {
-    let mut capacity = 0usize;
-    let mut requested = 0usize;
-    for length in 0..count {
-        if length == capacity {
-            capacity = grown_capacity(capacity, elem_size);
-            requested += capacity * elem_size;
-        }
-    }
-    (requested, capacity)
-}
-
-/// Fixed accounts on the host's `fhe_execute` CPI account struct (payer through the event CPI
-/// program). The invoke-measurement test in `heap_budget.rs` pins this against the real Anchor
-/// struct, so it cannot drift silently when the host's account list changes.
-pub const FHE_EXECUTE_FIXED_CPI_ACCOUNTS: usize = 9;
-
-/// Heap bytes the crate's own invoke path requests *after* `build()`: the CPI account-meta and
-/// account-info tables (`invoke`), and `resolve_accounts`'s three right-sized tables. An exact
-/// function of the account counts the builder already knows, charged against
-/// [`BUILD_HEAP_BUDGET_BYTES`] at `finish` — an execution admitted by `build()` fits the whole
-/// instruction, not just its own construction. Kept honest by the invoke measurement in
-/// `heap_budget.rs`, which runs the real assembly under a counting allocator.
-///
-/// Assumes the minimal invoke: zero per-transaction deny-subject witnesses (each deny record an
-/// app's transaction adds costs one more `AccountMeta` plus one more `AccountInfo` slot,
-/// ~90 bytes, out of the reserve), and every fixed account present — an absent optional HCU
-/// witness only makes the real cost smaller than charged.
-pub fn invoke_table_heap_bytes(
-    remaining_accounts: usize,
-    dynamic_accounts: usize,
-    output_authorities: usize,
-) -> usize {
-    let account_info_size = std::mem::size_of::<anchor_lang::prelude::AccountInfo<'static>>();
-    let account_meta_size =
-        std::mem::size_of::<anchor_lang::solana_program::instruction::AccountMeta>();
-    // Anchor's generated accessors grow the fixed accounts from `vec![]`; `invoke` then makes
-    // one exact reservation for the dynamic tail (no allocation when it still fits the grown
-    // capacity).
-    let table = |elem_size: usize| {
-        let (fixed_bytes, fixed_capacity) =
-            pushes_from_empty(FHE_EXECUTE_FIXED_CPI_ACCOUNTS, elem_size);
-        let final_length = FHE_EXECUTE_FIXED_CPI_ACCOUNTS + remaining_accounts;
-        let tail_bytes = if final_length > fixed_capacity {
-            final_length * elem_size
-        } else {
-            0
-        };
-        fixed_bytes + tail_bytes
-    };
-    // Anchor's generated `to_account_infos` builds each fixed account's contribution as a
-    // one-element `vec![clone]` before extending the table with it — one exact-size temporary
-    // per present field.
-    let fixed_info_temporaries = FHE_EXECUTE_FIXED_CPI_ACCOUNTS * account_info_size;
-    // `resolve_accounts` sizes its dynamic-account and authority-witness tables from the counts
-    // the execution itself requires, and its resolved table holds every remaining account.
-    let resolve_bytes =
-        (dynamic_accounts + output_authorities + remaining_accounts) * account_info_size;
-    table(account_meta_size) + table(account_info_size) + fixed_info_temporaries + resolve_bytes
-}
-
 /// The instructions a transaction is guaranteed to execute for one `fhe_execute` invocation in
 /// the minimal production wrapper: the app instruction, its `fhe_execute` CPI, three
 /// system-program CPIs per created persistent output, and one event CPI per event kind emitted
